services/auth/requirements.txt
flask==1.1.1
flask-cors==3.0.8
gunicorn==20.0.4
python-jose==3.1.0
requests==2.22.0